Adding more unit tests.
I've now covered all of the CLI entrypoints. A lot of the test aren't
great due to the DB dependency, but it's reasonable for now.

Coverage = 98%

The missing coverage is due to a bug in click that prevents one of my
tests from running: https://github.com/mitsuhiko/click/issues/344. I
wrote the test anyway and have marked it as a skip for now. We can
re-enable it later.