Change the way schemas are cleared.
This change is prompted by my first real-world usage of Agnostic today.
Dropping an entire database requires special privileges in Postgres,
and creating a database it not appropriate because you don't know which
template to use! It is much cleaner to drop all of the objects in the
database which are owned by the current user.