Replace psycopg2 driver with pg8000.
The latter is a pure Python driver. Psycopg2 doesn't publish Linux
wheels and it requires Postgres headers for installation -- making
it very unlikely that the average user will be able to simply `pip
install` it. Pg8000 requires some tweaks (it doesn't support named
parameters in SQL) but it is very easy to `pip install`.