Improve test output.
Test output was hard to read because we have 2 classes that implement
the same set of test methods, e.g.:

  root@60d25a25117a:/opt/agnostic# python3 -m tests -v
  The "bootstrap" CLI command creates a migrations table. ... ok
  The "bootstrap" CLI command does not recreate the migrations table. ... ok
  The "bootstrap" CLI command creates a migrations table. ... ok
  The "bootstrap" CLI command does not recreate the migrations table. ... ok

Note how the tests are repeated twice -- once for MySQL and once more for
Postgres -- but the reason for the repetition isn't clear from this output.
Furthermore, when a test fails, the traceback will show the failure occurred
in an abstract base class, making it difficult to ascertain which class
needs to be fixed.

This commit adds a Nose plugin (and a Nose wrapper that registers the plugin)
that prints the class name next to each test:

  root@60d25a25117a:/opt/agnostic# python3 -m tests --with-reportclassname -v
  [TestMysql] The "bootstrap" CLI command creates a migrations table. ... ok
  [TestMysql] The "bootstrap" CLI command does not recreate the migrations table. ... ok
  [TestPostgreSql] The "bootstrap" CLI command creates a migrations table. ... ok
  [TestPostgreSql] The "bootstrap" CLI command does not recreate the migrations table. ... ok

The new output makes it clear that two different classes are being tested, and
if a failure occurs, it will also be clear which class it occurred in.