Modify travis config to test inside Docker
Apparently Travis natively supports Docker now, although the docs
are very thin. It should be possible to have Travis run tests
inside the same exact Docker image that we are using for local
tests. Furthermore, this gives us better control over the external
dependencies, like Postgres 9.5 (not built in to Travis) or MySQL
5.7 (also not built in).

On the other hand, there's not enough docs for me to know if this
new Travis config will actually work or not, so I'm just crossing
my fingers and pushing what I have.